requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
Django==4.0.2
django-cors-headers==3.10.0
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.4.1
promise==2.3
PyJWT==1.7.1
pylint==2.12.2
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.0.1
wrapt==1.13.3
